bitbake.conf: change APACHE_MIRROR to point at archive.apache.org
The official download servers www.[country].apace.org only host the latest
release, so the URL is only valid when the recipe is fully up to date.

In the general case this isn't a problem as our mirror list includes
archive.apache.org, but the upstream URI checking (the checkuri task) fails as
that explicitly doesn't use the mirrors.

Signed-off-by: Ross Burton <ross.burton@intel.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>